+++ docs/CHANGELOG.md
@@ -43,6 +43,7 @@ Version 1.9.0 (not released yet)
 - Do not load unnecessary VPPs in dedicated server mode
 - Add level filename to "Level Initializing" console message
 - Add `map_rand` command and `vote rand` - both initiate a level change to a random level on the rotation
+- Add `$DF Randomize Rotation` option in dedicated server config
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------